+++ cscs_keygen/__init__.py
@@ -1,0 +1,3 @@
+from .cscs_keygen import entry_point as cli  # noqa: F401
+
+__version__ = "0.1.0"
