+++ cscs_keygen/__main__.py
@@ -1,0 +1,3 @@
+from .cscs_keygen import entry_point
+
+entry_point()
